+++ src/editor.js
@@ -99,7 +99,7 @@ OCA.Onlyoffice.InitEditor = function() {
 			}
 
 			const script = document.createElement('script')
-			script.src = config.documentServerUrl + 'web-apps/apps/api/documents/api.js?shardKey=' + config.document.key
+			script.src = config.documentServerUrl + 'web-apps/apps/api/documents/api.js?shardkey=' + config.document.key
 			script.setAttribute('nonce', btoa(getRequestToken()))
 			script.onerror = function() {
 				OCA.Onlyoffice.showMessage(t(OCA.Onlyoffice.AppName, 'ONLYOFFICE cannot be reached. Please contact admin'), 'error', { timeout: -1 })
@@ -153,7 +153,7 @@ OCA.Onlyoffice.InitEditor = function() {
 				}
 
 				if ((OCA.Onlyoffice.inframe && !OCA.Onlyoffice.shareToken)
-					|| (OCA.Onlyoffice.currentUser.uid)) {
+					|| (OCA.Onlyoffice.currentUser?.uid)) {
 					config.events.onRequestSaveAs = OCA.Onlyoffice.onRequestSaveAs
 					config.events.onRequestInsertImage = OCA.Onlyoffice.onRequestInsertImage
 					config.events.onRequestMailMergeRecipients = OCA.Onlyoffice.onRequestMailMergeRecipients
@@ -166,7 +166,7 @@ OCA.Onlyoffice.InitEditor = function() {
 					config.events.onMetaChange = OCA.Onlyoffice.onMetaChange
 					config.events.onRequestRefreshFile = OCA.Onlyoffice.onRequestRefreshFile
 
-					if (OCA.Onlyoffice.currentUser.uid) {
+					if (OCA.Onlyoffice.currentUser?.uid) {
 						config.events.onRequestUsers = OCA.Onlyoffice.onRequestUsers
 					}